+++ code/server/app/src/main/java/server/web/root/api/AccountAPI.java
@@ -27,7 +27,7 @@ import java.util.List;
 public class AccountAPI {
 
     public record UserInfo(
-            int id,
+            long id,
             String name,
             String email,
             String bio,
@@ -38,7 +38,7 @@ public class AccountAPI {
     @Route
     public static @Json UserInfo all_userinfo(@FromRequest(RequireSession.class) UserSession auth, RoConn conn) throws SQLException {
         try(var stmt = conn.namedPreparedStatement("select name, bio from users where id=:id")){
-            stmt.setInt(":id", auth.user_id);
+            stmt.setLong(":id", auth.user_id);
             var rs = stmt.executeQuery();
             return new UserInfo(
                     auth.user_id,
@@ -81,14 +81,14 @@ public class AccountAPI {
         cp.new_password = Util.hashy((cp.new_password+"\0\0\0\0"+cp.email).getBytes());
         try(var stmt = trans.namedPreparedStatement("update users set password=:new_password where email=:email AND password=:old_password AND id=:id")){
             stmt.setString(":email", cp.email);
-            stmt.setInt(":id", auth.user_id);
+            stmt.setLong(":id", auth.user_id);
             stmt.setString(":old_password", cp.old_password);
             stmt.setString(":new_password", cp.new_password);
             stmt.execute();
         }
 
         try(var stmt = trans.namedPreparedStatement("delete from sessions where user_id=:id")){
-            stmt.setInt(":id", auth.user_id);
+            stmt.setLong(":id", auth.user_id);
             stmt.execute();
         }
     }
@@ -128,7 +128,7 @@ public class AccountAPI {
 
     @Route
     public static String login(MailServer mail, @FromRequest(IpHandler.class)InetAddress ip, @FromRequest(UserAgentHandler.class)String agent, RwTransaction trans,  @Body @Json Login login) throws SQLException, ClientError.Unauthorized {
-        int user_id;
+        long user_id;
         login.password = Util.hashy((login.password+"\0\0\0\0"+login.email).getBytes());
         try(var stmt = trans.namedPreparedStatement("select id from users where email=:email AND pass=:pass")){
             stmt.setString(":email", login.email);
@@ -137,19 +137,19 @@ public class AccountAPI {
             if(!res.next())
                 throw new ClientError.Unauthorized("An account with the specified email does not exist, or the specified password is incorrect");
             try{
-                user_id = res.getInt(1);
+                user_id = res.getLong(1);
             }catch (SQLException ignore){
                 throw new ClientError.Unauthorized("An account with the specified email does not exist, or the specified password is incorrect");
             }
         }
 
-        int session_id;
+        long session_id;
         try(var stmt = trans.namedPreparedStatement("insert into sessions values(null, null, :user_id, :exp, :agent, :ip) returning id")){
-            stmt.setInt(":user_id", user_id);
+            stmt.setLong(":user_id", user_id);
             stmt.setLong(":exp", new Date().getTime() + 2628000000L);
             stmt.setString(":agent", agent);
             stmt.setString(":ip", ip.getHostAddress());
-            session_id = stmt.executeQuery().getInt(1);
+            session_id = stmt.executeQuery().getLong(1);
         }
 
         var hash = Util.hashy((login.email + "\0\0\0\0" + login.password + "\0\0\0\0" + session_id).getBytes());
@@ -157,7 +157,7 @@ public class AccountAPI {
 
         try(var stmt = trans.namedPreparedStatement("update sessions set token=:token where id=:id")){
             stmt.setString(":token", token);
-            stmt.setInt(":id", session_id);
+            stmt.setLong(":id", session_id);
             stmt.execute();
         }
 
@@ -176,7 +176,7 @@ public class AccountAPI {
     }
 
     public static class Session{
-        public int id;
+        public long id;
         public long expiration;
         public String agent;
         public String ip;
@@ -185,16 +185,16 @@ public class AccountAPI {
     @Route
     public static @Json List<Session> list_sessions(@FromRequest(RequireSession.class) UserSession auth, RoConn conn) throws SQLException {
         try(var stmt = conn.namedPreparedStatement("select * from sessions where user_id=:id")){
-            stmt.setInt(":id", auth.user_id);
+            stmt.setLong(":id", auth.user_id);
             return SqlSerde.sqlList(stmt.executeQuery(), Session.class);
         }
     }
 
     @Route("/invalidate_session/<session_id>")
-    public static void invalidate_session(@FromRequest(RequireSession.class) UserSession auth, RwConn conn, @Path int session_id) throws SQLException, ClientError.BadRequest {
+    public static void invalidate_session(@FromRequest(RequireSession.class) UserSession auth, RwConn conn, @Path long session_id) throws SQLException, ClientError.BadRequest {
         try(var stmt = conn.namedPreparedStatement("delete from sessions where id=:session_id AND user_id=:user_id")){
-            stmt.setInt(":session_id", session_id);
-            stmt.setInt(":user_id", auth.user_id);
+            stmt.setLong(":session_id", session_id);
+            stmt.setLong(":user_id", auth.user_id);
             if(stmt.executeUpdate() != 1)
                 throw new ClientError.BadRequest("Could not invalidate session, session does not belong to you or does not exist");
         }
